Creates natbib class option and updates text in the template
This verison changes the loading of the natbib package to an
option in the tamu_thesis class. As not all bibiography
styles require natbib, with those that don't generating
errors messages on compile, natbib is only loaded if
the option is set.

Much of the text in the first two sections of the document
has been changed to better explain how to use the template.
Includes better explaination of citations, acronym usage,
and cross-referencing.

Also add the \figref and \tabref commands so to make it
slightly easier to reference figures and tables.

v0.18.04c 20180515 1452 CDT