Moves nomenclature formatting to tamu_thesis.cls
This update removes the nomenclature.sty package as
all the required formatting is now contained in the
tamu_thesis.cls file. The all  and hover options for
the nomenclature package can be called on the
tamu_thesis class in the documentclass command.
Options have been renamed acroall and acrohover.

A changes.log file has also been added to highlight
changes and inform users which files are required for
changes to take effect.

v0.18.04d 20180619 1250 CDT