Moves to TeXLive 2020, bug fixes
There was a major issue with acronyms and running per section
reference lists. For more information, see this site:
  https://tex.stackexchange.com/questions/549782/compile-error-using-bibunits-and-acro-packages

To resolve this issue, moved package requirements forward to
TeXLive 2020 and acro v3, which fixes the issue.

Other main change is the addition of the phd and thesis options
for the tamu_thesis class. These options set the \Degree{} and
\PaperType{} tokens, respectively. This change is made to reduce
the amount of text the end-user must enter. The \Degree{} and
\PaperType{} tokens are still valid for backwards compatability
and will override any options set at class load.

Examples usage:
  If writing a thesis for your Master's degree
    \documentclass[12pt,thesis]{tamu_thesis}
  This will set \Degree{MASTER OF SCIENCE} and \PaperType{Thesis}

  For dissertation for PhD
    \documentclass[12pt,phd]{tamu_thesis}
  This will set \Degree{DOCTOR OF PHILOSOPHY} and \PaperType{Dissertation}

  Any combination of these keys can be used

Some other updates:
 - Updates to use texlive 2020.
 - Fixes issue with bibliography per section that occured due to
   acro v2 and bibunits. Now require acro v3 for this to work
   correctly
 - Updates acronym printing to work with acro v3
 - Adds the thesis and phd options, which set the PaperType and
   Degree, respectively. The \PaperType{} and \Degree{} options
   still exist, but better to use the thesis or phd options when
   loading the tamu_thesis class to eliminate spelling errors on
   the title page.
 - Cleans up the class file to remove old code

v0.20.06