Fixes empty List of Tables in ToC
Even if no tables are in the document, an empty List of Tables
was created at the beginning of the document along with the LoT
being added to the Table of Contents. This would also happen if
no figures were in the document, with an empty List of Figures
created. This has been fixed using the {totalcount} package and
using the \iftotalfigures and \iftotaltables commands provided
by that package. Now, the creation of the LoF and LoT is only
done if there are figures or tables in the document.

Note that the \makeNomenclature command had to also be modified as
it uses a table to create the nomenclature list. Thus, in the
\makeNomenclature command, the table counter is decremented by one
to ensure the nomenclature table is NOT counter toward the number
of tables in the document; i.e., \addtocounter{table}{-1}.

v0.20.08