Fixes citations in LoF/LoT when using chapter ref
There was an issue where citations in figure and table captions
would not appear in the List of Figures/Tables when using references
per chapter. This was solved using the {bibentry} package and the
\nobibliography} command, which enables references without printing
a bibliography section.

There was also a bug introduced in the last update that caused the
page numbering to be wrong; never changed to arabic from roman
after ToC, LoF, etc. This has been fixed.

NOTE: the main tex file has been updated as the \makebibliography
command is now always called. The command checks the \chapref flag
for log.

v0.20.08a