Adds ability to adjust chapter ToC entries
The ability to adjust ToC entries for chapters has been re-enabled
through keywords. This functionality was removed in a pervious
version when the ability to 'easily' add footnotes was added.
However, now users can use the footnote= keyword to set a footnote
for the chapter and/or the toc= keyword to adjust how the chapter
title will appear in the ToC; can add extra line breaks if leaders
are missing.

v0.21.03a_tex2020